+++ x-pack/plugins/logs_shared/public/components/logging/log_text_stream/index.ts
@@ -8,11 +8,12 @@
 export type { LogEntryStreamItem } from './item';
 export type { LogEntryColumnWidths } from './log_entry_column';
 
-export { LogColumnHeader, LogColumnHeadersWrapper } from './column_headers';
+export { LogColumnHeader } from './column_headers';
+export { LogColumnHeadersWrapper } from './column_headers_wrapper';
 export { iconColumnId, LogEntryColumn, useColumnWidths } from './log_entry_column';
 export { LogEntryContextMenu } from './log_entry_context_menu';
 export { LogEntryFieldColumn } from './log_entry_field_column';
 export { LogEntryMessageColumn } from './log_entry_message_column';
-export { LogEntryRowWrapper } from './log_entry_row';
+export { LogEntryRowWrapper } from './log_entry_row_wrapper';
 export { LogEntryTimestampColumn } from './log_entry_timestamp_column';
 export { ScrollableLogTextStreamView } from './scrollable_log_text_stream_view';
