+++ x-pack/plugins/logs_shared/public/components/logging/log_text_stream/log_entry_column.tsx
@@ -156,3 +156,6 @@ export const useColumnWidths = ({
     [columnWidths, CharacterDimensionsProbe]
   );
 };
+
+// eslint-disable-next-line import/no-default-export
+export default LogEntryColumn;
