+++ x-pack/plugins/logs_shared/public/components/logging/log_text_stream/log_entry_field_column.tsx
@@ -64,3 +64,6 @@ const FieldColumnContent = euiStyled(LogEntryColumnContent)<LogEntryColumnConten
       ? preWrappedContentStyle
       : unwrappedContentStyle};
 `;
+
+// eslint-disable-next-line import/no-default-export
+export default LogEntryFieldColumn;
