+++ x-pack/plugins/logs_shared/public/components/logging/log_text_stream/log_entry_message_column.tsx
@@ -114,3 +114,6 @@ const renderMessageSegments = (messageSegments: LogMessagePart[]): string => {
     })
     .join(' ');
 };
+
+// eslint-disable-next-line import/no-default-export
+export default LogEntryMessageColumn;
