+++ x-pack/plugins/logs_shared/public/components/logging/log_text_stream/log_entry_timestamp_column.tsx
@@ -31,3 +31,6 @@ const TimestampColumnContent = euiStyled(LogEntryColumnContent)`
   text-overflow: clip;
   white-space: pre;
 `;
+
+// eslint-disable-next-line import/no-default-export
+export default LogEntryTimestampColumn;
